[MNT] deprecation of `proba` module in favour of `skpro` soft dependency (#6940)
The `proba` module has been merged into `skpro`, with which it is now
redundant.

This PR is the start of a transition where the probability distirbutions
module `proba` is merged into `skpro`:

* if `skpro` is present, imports from `proba` will be redirected, until
0.35.0.
* if not, the present module is used for imports, and an informative
warning is raised about deprecation.

The deprecation schedule is:

* 0.32.0 - informative warnings, dispatch as above
* 0.35.0 - removal of local logic. informative errors in case `skpro` is
not present
* 0.38.0 - complete removal of `proba` module. Probability distribution
logic uses `skpro` directly, and soft dependency errors are raised if
not present.

This should give users enought time to transition - the required action
is simple, adding `skpro` to the python environment if not already
present, in cases of users consuming the probabilistic forecasting
interfaces, only those relying on distributional predictions from
`predict_proba` (this does not impact `predict_interval` etc).
