+++ sktime/proba/__init__.py
@@ -1,1 +1,7 @@
 """Probability distribution objects."""
+
+# TODO 0.35.0 - remove sktime logic for distributions
+# replace imports from sktime with informative error message
+# replace imports from proba module within sktime with imports from skpro
+#
+# TODO 0.38.0 - remove sktime proba module
