+++ sktime/utils/_testing/scenarios_getter.py
@@ -7,7 +7,8 @@ __all__ = ["retrieve_scenarios"]
 
 from inspect import isclass
 
-from sktime.base import BaseObject
+from skbase.base import BaseObject
+
 from sktime.registry import BASE_CLASS_LIST, BASE_CLASS_SCITYPE_LIST, scitype
 from sktime.utils._testing.scenarios_aligners import scenarios_aligners
 from sktime.utils._testing.scenarios_classification import (
